PR comment updates
