+++ .gitignore
@@ -91,4 +91,3 @@ tags
 
 # IDE
 .vscode/
-.idea/
